pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "epiminds-swarm"
version = "0.1.0"
description = "Swarm agent architecture boilerplate (FastAPI + Gemini)."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Epiminds" }]
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "google-genai>=0.7",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "ruff>=0.4",
  "httpx>=0.27",
]

[tool.hatch.build.targets.wheel]
packages = ["src/swarm"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

